Refactor code to be shared between multiple endpoint
The operator endpoint needs almost exactly the same viewsets that
are implemented for enforcement endpoint. Therefore move these
existing viewsets to global level, so that it can be shared between
operator and enforcement endpoints.
